DEV-12851 Further refinement and helper methods
