+++ index.js
@@ -11,7 +11,8 @@ import {
     buildFoundationsApiResponseOptions,
     returnDataIfSuccessfulOrThrowError, throwUnexpectedResponseCodeError } from './api-services/helpers'
 import { buildProtectiveMonitoringOptions } from './services/helpers'
-import { isHealthCheckRequest, isResourceRequest, logRequestInfo } from './common/hapi-request.helpers'
+import { isHealthCheckRequest, isResourceRequest } from './common/hapi-request.helpers'
+import { logRequestInfo, logRequestError } from './common/logging.helpers'
 import { isHealthUrl, isResourceUrl } from './common/url-string.helpers'
 
 export {
@@ -38,5 +39,7 @@ export {
 
     isHealthCheckRequest, isResourceRequest,
 
-    isHealthUrl, isResourceUrl, logRequestInfo
+    isHealthUrl, isResourceUrl,
+
+    logRequestInfo, logRequestError
 }
